+++ src/redux/otomiApi.ts
@@ -116,6 +116,9 @@ const injectedRtkApi = api.injectEndpoints({
         body: queryArg.body,
       }),
     }),
+    getK8SVersion: build.query<GetK8SVersionApiResponse, GetK8SVersionApiArg>({
+      query: () => ({ url: `/k8sVersion` }),
+    }),
     connectCloudtty: build.mutation<ConnectCloudttyApiResponse, ConnectCloudttyApiArg>({
       query: (queryArg) => ({ url: `/cloudtty`, method: 'POST', body: queryArg.body }),
     }),
@@ -296,9 +299,14 @@ export type GetAllServicesApiResponse = /** status 200 Successfully obtained all
         | ({
             ingressClassName?: string
             tlsPass?: boolean
-            useDefaultSubdomain?: boolean
+            useDefaultHost?: boolean
             subdomain: string
             domain: string
+            useCname?: boolean
+            cname?: {
+              domain?: string
+              tlsSecretName?: string
+            }
             paths?: string[]
             forwardPath?: boolean
             hasCert?: boolean
@@ -948,9 +956,14 @@ export type GetTeamServicesApiResponse = /** status 200 Successfully obtained se
         | ({
             ingressClassName?: string
             tlsPass?: boolean
-            useDefaultSubdomain?: boolean
+            useDefaultHost?: boolean
             subdomain: string
             domain: string
+            useCname?: boolean
+            cname?: {
+              domain?: string
+              tlsSecretName?: string
+            }
             paths?: string[]
             forwardPath?: boolean
             hasCert?: boolean
@@ -1015,9 +1028,14 @@ export type CreateServiceApiResponse = /** status 200 Successfully stored servic
         | ({
             ingressClassName?: string
             tlsPass?: boolean
-            useDefaultSubdomain?: boolean
+            useDefaultHost?: boolean
             subdomain: string
             domain: string
+            useCname?: boolean
+            cname?: {
+              domain?: string
+              tlsSecretName?: string
+            }
             paths?: string[]
             forwardPath?: boolean
             hasCert?: boolean
@@ -1082,9 +1100,14 @@ export type CreateServiceApiArg = {
           | ({
               ingressClassName?: string
               tlsPass?: boolean
-              useDefaultSubdomain?: boolean
+              useDefaultHost?: boolean
               subdomain: string
               domain: string
+              useCname?: boolean
+              cname?: {
+                domain?: string
+                tlsSecretName?: string
+              }
               paths?: string[]
               forwardPath?: boolean
               hasCert?: boolean
@@ -1155,9 +1178,14 @@ export type GetServiceApiResponse = /** status 200 Successfully obtained service
         | ({
             ingressClassName?: string
             tlsPass?: boolean
-            useDefaultSubdomain?: boolean
+            useDefaultHost?: boolean
             subdomain: string
             domain: string
+            useCname?: boolean
+            cname?: {
+              domain?: string
+              tlsSecretName?: string
+            }
             paths?: string[]
             forwardPath?: boolean
             hasCert?: boolean
@@ -1224,9 +1252,14 @@ export type EditServiceApiResponse = /** status 200 Successfully edited service 
         | ({
             ingressClassName?: string
             tlsPass?: boolean
-            useDefaultSubdomain?: boolean
+            useDefaultHost?: boolean
             subdomain: string
             domain: string
+            useCname?: boolean
+            cname?: {
+              domain?: string
+              tlsSecretName?: string
+            }
             paths?: string[]
             forwardPath?: boolean
             hasCert?: boolean
@@ -1293,9 +1326,14 @@ export type EditServiceApiArg = {
           | ({
               ingressClassName?: string
               tlsPass?: boolean
-              useDefaultSubdomain?: boolean
+              useDefaultHost?: boolean
               subdomain: string
               domain: string
+              useCname?: boolean
+              cname?: {
+                domain?: string
+                tlsSecretName?: string
+              }
               paths?: string[]
               forwardPath?: boolean
               hasCert?: boolean
@@ -1626,9 +1664,6 @@ export type GetAllBuildsApiResponse = /** status 200 Successfully obtained all b
         }
         type: 'buildpacks'
       }
-  repoAccess?: {
-    otomiGit?: boolean
-  }
 }[]
 export type GetAllBuildsApiArg = void
 export type GetTeamBuildsApiResponse = /** status 200 Successfully obtained team builds configuration */ {
@@ -1653,9 +1688,6 @@ export type GetTeamBuildsApiResponse = /** status 200 Successfully obtained team
         }
         type: 'buildpacks'
       }
-  repoAccess?: {
-    otomiGit?: boolean
-  }
 }[]
 export type GetTeamBuildsApiArg = {
   /** ID of team to return */
@@ -1683,9 +1715,6 @@ export type CreateBuildApiResponse = /** status 200 Successfully stored build co
         }
         type: 'buildpacks'
       }
-  repoAccess?: {
-    otomiGit?: boolean
-  }
 }
 export type CreateBuildApiArg = {
   /** ID of team to return */
@@ -1713,9 +1742,6 @@ export type CreateBuildApiArg = {
           }
           type: 'buildpacks'
         }
-    repoAccess?: {
-      otomiGit?: boolean
-    }
   }
 }
 export type DeleteBuildApiResponse = /** status 200 Successfully deleted a build */ undefined
@@ -1747,9 +1773,6 @@ export type GetBuildApiResponse = /** status 200 Successfully obtained build con
         }
         type: 'buildpacks'
       }
-  repoAccess?: {
-    otomiGit?: boolean
-  }
 }
 export type GetBuildApiArg = {
   /** ID of team to return */
@@ -1779,9 +1802,6 @@ export type EditBuildApiResponse = /** status 200 Successfully edited a team bui
         }
         type: 'buildpacks'
       }
-  repoAccess?: {
-    otomiGit?: boolean
-  }
 }
 export type EditBuildApiArg = {
   /** ID of team to return */
@@ -1811,11 +1831,10 @@ export type EditBuildApiArg = {
           }
           type: 'buildpacks'
         }
-    repoAccess?: {
-      otomiGit?: boolean
-    }
   }
 }
+export type GetK8SVersionApiResponse = /** status 200 Successfully obtained k8s version */ string
+export type GetK8SVersionApiArg = void
 export type ConnectCloudttyApiResponse = /** status 200 Successfully stored cloudtty configuration */ {
   id?: string
   teamId: string
@@ -1879,9 +1898,6 @@ export type GetAllProjectsApiResponse = /** status 200 Successfully obtained all
           }
           type: 'buildpacks'
         }
-    repoAccess?: {
-      otomiGit?: boolean
-    }
   }
   workload?: {
     id?: string
@@ -1990,9 +2006,14 @@ export type GetAllProjectsApiResponse = /** status 200 Successfully obtained all
           | ({
               ingressClassName?: string
               tlsPass?: boolean
-              useDefaultSubdomain?: boolean
+              useDefaultHost?: boolean
               subdomain: string
               domain: string
+              useCname?: boolean
+              cname?: {
+                domain?: string
+                tlsSecretName?: string
+              }
               paths?: string[]
               forwardPath?: boolean
               hasCert?: boolean
@@ -2064,9 +2085,6 @@ export type GetTeamProjectsApiResponse = /** status 200 Successfully obtained te
           }
           type: 'buildpacks'
         }
-    repoAccess?: {
-      otomiGit?: boolean
-    }
   }
   workload?: {
     id?: string
@@ -2175,9 +2193,14 @@ export type GetTeamProjectsApiResponse = /** status 200 Successfully obtained te
           | ({
               ingressClassName?: string
               tlsPass?: boolean
-              useDefaultSubdomain?: boolean
+              useDefaultHost?: boolean
               subdomain: string
               domain: string
+              useCname?: boolean
+              cname?: {
+                domain?: string
+                tlsSecretName?: string
+              }
               paths?: string[]
               forwardPath?: boolean
               hasCert?: boolean
@@ -2252,9 +2275,6 @@ export type CreateProjectApiResponse = /** status 200 Successfully stored projec
           }
           type: 'buildpacks'
         }
-    repoAccess?: {
-      otomiGit?: boolean
-    }
   }
   workload?: {
     id?: string
@@ -2363,9 +2383,14 @@ export type CreateProjectApiResponse = /** status 200 Successfully stored projec
           | ({
               ingressClassName?: string
               tlsPass?: boolean
-              useDefaultSubdomain?: boolean
+              useDefaultHost?: boolean
               subdomain: string
               domain: string
+              useCname?: boolean
+              cname?: {
+                domain?: string
+                tlsSecretName?: string
+              }
               paths?: string[]
               forwardPath?: boolean
               hasCert?: boolean
@@ -2440,9 +2465,6 @@ export type CreateProjectApiArg = {
             }
             type: 'buildpacks'
           }
-      repoAccess?: {
-        otomiGit?: boolean
-      }
     }
     workload?: {
       id?: string
@@ -2551,9 +2573,14 @@ export type CreateProjectApiArg = {
             | ({
                 ingressClassName?: string
                 tlsPass?: boolean
-                useDefaultSubdomain?: boolean
+                useDefaultHost?: boolean
                 subdomain: string
                 domain: string
+                useCname?: boolean
+                cname?: {
+                  domain?: string
+                  tlsSecretName?: string
+                }
                 paths?: string[]
                 forwardPath?: boolean
                 hasCert?: boolean
@@ -2632,9 +2659,6 @@ export type GetProjectApiResponse = /** status 200 Successfully obtained project
           }
           type: 'buildpacks'
         }
-    repoAccess?: {
-      otomiGit?: boolean
-    }
   }
   workload?: {
     id?: string
@@ -2743,9 +2767,14 @@ export type GetProjectApiResponse = /** status 200 Successfully obtained project
           | ({
               ingressClassName?: string
               tlsPass?: boolean
-              useDefaultSubdomain?: boolean
+              useDefaultHost?: boolean
               subdomain: string
               domain: string
+              useCname?: boolean
+              cname?: {
+                domain?: string
+                tlsSecretName?: string
+              }
               paths?: string[]
               forwardPath?: boolean
               hasCert?: boolean
@@ -2822,9 +2851,6 @@ export type EditProjectApiResponse = /** status 200 Successfully edited a team p
           }
           type: 'buildpacks'
         }
-    repoAccess?: {
-      otomiGit?: boolean
-    }
   }
   workload?: {
     id?: string
@@ -2933,9 +2959,14 @@ export type EditProjectApiResponse = /** status 200 Successfully edited a team p
           | ({
               ingressClassName?: string
               tlsPass?: boolean
-              useDefaultSubdomain?: boolean
+              useDefaultHost?: boolean
               subdomain: string
               domain: string
+              useCname?: boolean
+              cname?: {
+                domain?: string
+                tlsSecretName?: string
+              }
               paths?: string[]
               forwardPath?: boolean
               hasCert?: boolean
@@ -3012,9 +3043,6 @@ export type EditProjectApiArg = {
             }
             type: 'buildpacks'
           }
-      repoAccess?: {
-        otomiGit?: boolean
-      }
     }
     workload?: {
       id?: string
@@ -3123,9 +3151,14 @@ export type EditProjectApiArg = {
             | ({
                 ingressClassName?: string
                 tlsPass?: boolean
-                useDefaultSubdomain?: boolean
+                useDefaultHost?: boolean
                 subdomain: string
                 domain: string
+                useCname?: boolean
+                cname?: {
+                  domain?: string
+                  tlsSecretName?: string
+                }
                 paths?: string[]
                 forwardPath?: boolean
                 hasCert?: boolean
@@ -3797,7 +3830,6 @@ export type GetSettingsApiResponse = /** status 200 The request is successful. *
     name?: string
     domainSuffix?: string
     provider?: 'aws' | 'azure' | 'digitalocean' | 'google' | 'ovh' | 'vultr' | 'custom'
-    k8sVersion?: '1.23' | '1.24' | '1.25'
     apiName?: string
     apiServer?: string
     owner?: string
@@ -4222,7 +4254,6 @@ export type EditSettingsApiArg = {
       name?: string
       domainSuffix?: string
       provider?: 'aws' | 'azure' | 'digitalocean' | 'google' | 'ovh' | 'vultr' | 'custom'
-      k8sVersion?: '1.23' | '1.24' | '1.25'
       apiName?: string
       apiServer?: string
       owner?: string
@@ -4686,6 +4717,7 @@ export const {
   useDeleteBuildMutation,
   useGetBuildQuery,
   useEditBuildMutation,
+  useGetK8SVersionQuery,
   useConnectCloudttyMutation,
   useDeleteCloudttyMutation,
   useGetAllProjectsQuery,